style: unify into unchecked_calls.rs file
